UIDATIMP-1383: Move validator to utils
